pyproject.toml
[project]
name = "autotest-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pathspec>=1.0.4",
    "python-dotenv>=1.2.1",
]
